+++ Sources/reactant/Scaffold/Assets.swift
@@ -1,0 +1,87 @@
+//
+//  Assets.swift
+//  ReactantCLI
+//
+//  Created by Matous Hybl on 6/5/17.
+//
+//
+
+import Foundation
+func assetsJson() -> [String] {
+    return [
+        "{",
+        "    \"info\" : {",
+        "        \"version\" : 1,",
+        "        \"author\" : \"xcode\"",
+        "    }",
+        "}",
+    ]
+}
+
+func appIconJson() -> [String] {
+    return [
+        "{",
+        "  \"images\" : [",
+        "    {",
+        "      \"idiom\" : \"iphone\",",
+        "      \"size\" : \"20x20\",",
+        "      \"scale\" : \"2x\"",
+        "    },",
+        "    {",
+        "      \"idiom\" : \"iphone\",",
+        "      \"size\" : \"20x20\",",
+        "      \"scale\" : \"3x\"",
+        "    },",
+        "    {",
+        "      \"idiom\" : \"iphone\",",
+        "      \"size\" : \"29x29\",",
+        "      \"scale\" : \"1x\"",
+        "    },",
+        "    {",
+        "      \"idiom\" : \"iphone\",",
+        "      \"size\" : \"29x29\",",
+        "      \"scale\" : \"2x\"",
+        "    },",
+        "    {",
+        "      \"idiom\" : \"iphone\",",
+        "      \"size\" : \"29x29\",",
+        "      \"scale\" : \"3x\"",
+        "    },",
+        "    {",
+        "      \"idiom\" : \"iphone\",",
+        "      \"size\" : \"40x40\",",
+        "      \"scale\" : \"2x\"",
+        "    },",
+        "    {",
+        "      \"idiom\" : \"iphone\",",
+        "      \"size\" : \"40x40\",",
+        "      \"scale\" : \"3x\"",
+        "    },",
+        "    {",
+        "      \"idiom\" : \"iphone\",",
+        "      \"size\" : \"57x57\",",
+        "      \"scale\" : \"1x\"",
+        "    },",
+        "    {",
+        "      \"idiom\" : \"iphone\",",
+        "      \"size\" : \"57x57\",",
+        "      \"scale\" : \"2x\"",
+        "    },",
+        "    {",
+        "      \"idiom\" : \"iphone\",",
+        "      \"size\" : \"60x60\",",
+        "      \"scale\" : \"2x\"",
+        "    },",
+        "    {",
+        "      \"idiom\" : \"iphone\",",
+        "      \"size\" : \"60x60\",",
+        "      \"scale\" : \"3x\"",
+        "    }",
+        "  ],",
+        "  \"info\" : {",
+        "    \"version\" : 1,",
+        "    \"author\" : \"xcode\"",
+        "  }",
+        "}",
+    ]
+}
