+++ Sources/reactant/Scaffold/LaunchScreen.swift
@@ -1,0 +1,41 @@
+//
+//  LaunchScreen.swift
+//  ReactantCLI
+//
+//  Created by Matous Hybl on 6/5/17.
+//
+//
+
+import Foundation
+
+func launchScreen() -> [String] {
+    return [
+        "<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"no\"?>",
+        "<document type=\"com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB\" version=\"3.0\" toolsVersion=\"11134\" systemVersion=\"15F34\" targetRuntime=\"iOS.CocoaTouch\" propertyAccessControl=\"none\" useAutolayout=\"YES\"", "launchScreen=\"YES\" useTraitCollections=\"YES\" colorMatched=\"YES\" initialViewController=\"01J-lp-oVM\">",
+        "    <dependencies>",
+        "        <plugIn identifier=\"com.apple.InterfaceBuilder.IBCocoaTouchPlugin\" version=\"11106\"/>",
+        "        <capability name=\"documents saved in the Xcode 8 format\" minToolsVersion=\"8.0\"/>",
+        "    </dependencies>",
+        "    <scenes>",
+        "        <!--View Controller-->",
+        "        <scene sceneID=\"EHf-IW-A2E\">",
+        "            <objects>",
+        "                <viewController id=\"01J-lp-oVM\" sceneMemberID=\"viewController\">",
+        "                    <layoutGuides>",
+        "                        <viewControllerLayoutGuide type=\"top\" id=\"Llm-lL-Icb\"/>",
+        "                        <viewControllerLayoutGuide type=\"bottom\" id=\"xb3-aO-Qok\"/>",
+        "                    </layoutGuides>",
+        "                    <view key=\"view\" contentMode=\"scaleToFill\" id=\"Ze5-6b-2t3\">",
+        "                        <rect key=\"frame\" x=\"0.0\" y=\"0.0\" width=\"375\" height=\"667\"/>",
+        "                        <autoresizingMask key=\"autoresizingMask\" widthSizable=\"YES\" heightSizable=\"YES\"/>",
+        "                        <color key=\"backgroundColor\" red=\"1\" green=\"1\" blue=\"1\" alpha=\"1\" colorSpace=\"custom\" customColorSpace=\"sRGB\"/>",
+        "                    </view>",
+        "                </viewController>",
+        "                <placeholder placeholderIdentifier=\"IBFirstResponder\" id=\"iYj-Kq-Ea1\" userLabel=\"First Responder\" sceneMemberID=\"firstResponder\"/>",
+        "            </objects>",
+        "            <point key=\"canvasLocation\" x=\"53\" y=\"375\"/>",
+        "        </scene>",
+        "    </scenes>",
+        "</document>",
+    ]
+}
