+++ Sources/reactant/Scaffold/Service.swift
@@ -1,0 +1,17 @@
+//
+//  Service.swift
+//  ReactantCLI
+//
+//  Created by Matous Hybl on 6/5/17.
+//
+//
+
+import Foundation
+
+func sampleService() -> [String] {
+    return [
+        "import Foundation",
+        "",
+        "struct SampleService { }",
+    ]
+}
