+++ Sources/reactant/VersionControl.swift
@@ -1,0 +1,14 @@
+//
+//  VersionControl.swift
+//  ReactantCLI
+//
+//  Created by Matous Hybl on 6/5/17.
+//
+//
+
+import Foundation
+
+enum VersionControl: String {
+    case git
+    case none
+}
